Random pass for a few style issues.
- Standardize on using `New()` vs `NewCirconusMetrics()`
- Move most of the public tests to a dedicated `_test` package to test
  the public interface of the package.
- Stand up an `-impl_test.go` file to peak behind the curtains and test
  the implementation.
